+++ android/fastlane/report.xml
@@ -5,17 +5,17 @@
     
     
       
-      <testcase classname="fastlane.lanes" name="0: default_platform" time="0.00015">
+      <testcase classname="fastlane.lanes" name="0: default_platform" time="0.000191">
         
       </testcase>
     
       
-      <testcase classname="fastlane.lanes" name="1: flutter build apk --release --target-platform=android-arm64" time="4.385869">
+      <testcase classname="fastlane.lanes" name="1: flutter build apk --release --target-platform=android-arm64" time="45.38843">
         
       </testcase>
     
       
-      <testcase classname="fastlane.lanes" name="2: firebase_app_distribution" time="11.844974">
+      <testcase classname="fastlane.lanes" name="2: firebase_app_distribution" time="13.770247">
         
       </testcase>
