+++ lib/view/club_schedule/components/calendar/club_schedule_table_calendar_month.dart
@@ -76,6 +76,7 @@ class ClubScheduleTableCalendarMonth extends ConsumerWidget {
               padding: const EdgeInsets.only(
                 left: defaultPaddingS,
                 right: defaultPaddingS,
+                bottom: defaultPaddingS / 2,
               ),
               child: Row(
                 children: [
