+++ lib/view/group/group_list_page.dart
@@ -2,6 +2,7 @@ import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:material_symbols_icons/symbols.dart';
+import 'package:share_plus/share_plus.dart';
 import 'package:woohakdong/model/group/group.dart';
 import 'package:woohakdong/view/group/components/group_list_tile.dart';
 import 'package:woohakdong/view/group/group_detail_page.dart';
@@ -11,8 +12,12 @@ import 'package:woohakdong/view/themes/theme_context.dart';
 import 'package:woohakdong/view_model/group/group_list_provider.dart';
 import 'package:woohakdong/view_model/group/group_provider.dart';
 
+import '../../service/general/general_format.dart';
+import '../../service/general/general_functions.dart';
+import '../themes/custom_widget/dialog/custom_interaction_dialog.dart';
 import '../themes/custom_widget/etc/custom_horizontal_divider.dart';
 import 'group_add_page.dart';
+import 'group_edit_page.dart';
 
 class GroupListPage extends ConsumerWidget {
   const GroupListPage({super.key});
@@ -52,6 +57,9 @@ class GroupListPage extends ConsumerWidget {
                 itemBuilder: (context, index) => GroupListTile(
                   group: groupList[index],
                   onTap: () => _pushGroupDetailPage(ref, context, groupList[index].groupId!),
+                  onEditLongPress: () => _pushGroupEditPage(context, groupList[index]),
+                  onDeleteLongPress: () async => await _deleteGroup(context, ref, groupList[index]),
+                  onShareLongPress: () => _onShareTap(groupList[index]),
                 ),
               ),
             );
@@ -105,6 +113,44 @@ class GroupListPage extends ConsumerWidget {
           builder: (context) => const GroupDetailPage(),
         ),
       );
+    }
+  }
+
+  void _onShareTap(Group groupInfo) {
+    Share.share(
+      groupInfo.groupJoinLink!,
+      subject: '${groupInfo.groupName}에서 만나요! 🤗\n\n'
+          '모임비: ${GeneralFormat.formatClubDues(groupInfo.groupAmount!)}\n'
+          '최대 인원: ${groupInfo.groupMemberLimit}명\n'
+          '현재 인원: ${groupInfo.groupMemberCount}명\n\n',
+    );
+  }
+
+  void _pushGroupEditPage(BuildContext context, Group groupInfo) {
+    Navigator.push(
+      context,
+      CupertinoPageRoute(
+        builder: (context) => GroupEditPage(groupInfo: groupInfo),
+      ),
+    );
+  }
+
+  Future<void> _deleteGroup(BuildContext context, WidgetRef ref, Group groupInfo) async {
+    try {
+      final bool? isDelete = await showDialog<bool>(
+        context: context,
+        builder: (context) => const CustomInteractionDialog(
+          dialogTitle: '모임 삭제',
+          dialogContent: '모임을 삭제하면 되돌릴 수 없어요.',
+        ),
+      );
+
+      if (isDelete != true) return;
+
+      await ref.read(groupProvider.notifier).deleteGroup(groupInfo.groupId!);
+      GeneralFunctions.toastMessage('모임이 삭제되었어요');
+    } catch (e) {
+      GeneralFunctions.toastMessage('오류가 발생했어요\n다시 시도해 주세요');
     }
   }
 }
